pyproject.toml
[tool.poetry]
name = "hotel-booking-project"
version = "0.1.0"
description = ""
authors = ["Saveliy <s.lekhtin@innopolis.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
black = "^23.11.0"
isort = "^5.12.0"
psycopg2 = "^2.9.9"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
